wip: refactoring tests
