+++ test/querySelector.spec.js
@@ -6,7 +6,8 @@
  */
 /* global it describe */
 
-import { querySelectorAll, querySelector, getElementsByTagName, getElementsByTagNameNS } from '../src/index.js'
+import { querySelectorAll, querySelector } from '../src/index.js'
+import { assertResultEqual, withDom } from './utils'
 import assert from 'assert'
 import simpleLight1 from './fixtures/simple1/light.html'
 import simpleShadow1 from './fixtures/simple1/shadow.html'
@@ -46,54 +47,9 @@ import nestedSlotsLight6 from './fixtures/nestedSlots6/light.html'
 import nestedSlotsShadow6 from './fixtures/nestedSlots6/shadow.html'
 import nestedSlotsLight7 from './fixtures/nestedSlots7/light.html'
 import nestedSlotsShadow7 from './fixtures/nestedSlots7/shadow.html'
-import tagNameLight1 from './fixtures/tagName1/light.html'
-import tagNameShadow1 from './fixtures/tagName1/shadow.html'
 
-function withDom (html, cb) {
-  const iframe = document.createElement('iframe')
-  document.body.appendChild(iframe)
-  const iframeDocument = iframe.contentWindow.document
-  iframeDocument.open('text/html', 'replace')
-  iframeDocument.write(html)
-  iframeDocument.close()
-  try {
-    cb(iframeDocument)
-  } finally {
-    document.body.removeChild(iframe)
-  }
-}
 
-function simplifyElement (element) {
-  if (!element) {
-    return undefined
-  }
-  return {
-    tagName: element.tagName,
-    classList: [...element.classList]
-  }
-}
 
-function simplifyElements (elements) {
-  return [...elements].map(simplifyElement)
-}
-
-function stringify (obj) {
-  if (typeof obj === 'undefined') {
-    return obj
-  }
-  return JSON.stringify(obj)
-}
-
-function assertResultEqual (selector, actual, expected, qsa) {
-  if (qsa) {
-    actual = simplifyElements(actual)
-  } else {
-    actual = simplifyElement(actual)
-    expected = expected[0]
-  }
-  assert.deepStrictEqual(actual, expected,
-    `Selector failed: ${stringify(selector)}, ${stringify(actual)} !== ${stringify(expected)}`)
-}
 
 function testSelectors (lightDom, shadowDom, tests) {
   tests.forEach(({ selector, expected }) => {
@@ -893,397 +849,5 @@ describe('basic test suite', function () {
 
     withDom(simpleLight1, test(true))
     withDom(simpleShadow1, test(false))
-  })
-
-  describe('getElementsByTagName', () => {
-    it('light DOM - handles tag names', () => {
-      const tagName1 = "svg";
-      const expected1 = [
-        {
-          tagName: 'svg',
-          classList: []
-        }
-      ]
-      withDom(tagNameLight1, context => {
-        assertResultEqual(tagName1, context.getElementsByTagName(tagName1), expected1, true)
-      })
-      const tagName2 = "span";
-      const expected2 = [
-        {
-          tagName: 'SPAN',
-          classList: []
-        }
-      ]
-      withDom(tagNameLight1, context => {
-        assertResultEqual(tagName2, context.getElementsByTagName(tagName2), expected2, true)
-      })
-    });
-
-    it('shadow DOM - handles tag names', () => {
-      const tagName1 = "svg";
-      const expected1 = [
-        {
-          tagName: 'svg',
-          classList: []
-        }
-      ]
-      withDom(tagNameShadow1, context => {
-        assertResultEqual(tagName1, getElementsByTagName(tagName1, context), expected1, true)
-      })
-      const tagName2 = "span";
-      const expected2 = [
-        {
-          tagName: 'SPAN',
-          classList: []
-        }
-      ]
-      withDom(tagNameShadow1, context => {
-        assertResultEqual(tagName2, getElementsByTagName(tagName2, context), expected2, true)
-      })
-    });
-
-    it('light DOM - handles wildcards', () => {
-      const tagName = "*";
-      const expected = [
-        {
-          tagName: 'HTML',
-          classList: []
-        },
-        {
-          tagName: 'HEAD',
-          classList: []
-        },
-        {
-          tagName: 'BODY',
-          classList: []
-        },
-        {
-          tagName: 'DIV',
-          classList: ['container']
-        },
-        {
-          tagName: 'DIV',
-          classList: ['outer-component']
-        },
-        {
-          tagName: 'DIV',
-          classList: ['wrapper']
-        },
-        {
-          tagName: 'DIV',
-          classList: ['inner-component']
-        },
-        {
-          tagName: 'SPAN',
-          classList: []
-        },
-        {
-          tagName: 'svg',
-          classList: []
-        },
-        {
-          tagName: 'defs',
-          classList: []
-        },
-        {
-          tagName: 'linearGradient',
-          classList: []
-        },
-        {
-          tagName: 'stop',
-          classList: []
-        },
-        {
-          tagName: 'stop',
-          classList: []
-        },
-        {
-          tagName: 'circle',
-          classList: []
-        }
-      ]
-      withDom(tagNameLight1, context => {
-        assertResultEqual(tagName, context.getElementsByTagName(tagName), expected, true)
-      })
-    });
-
-    it('shadow DOM - handles wildcards', () => {
-      const tagName = "*";
-      const expected = [
-        {
-          tagName: 'HEAD',
-          classList: []
-        },
-        {
-          tagName: 'BODY',
-          classList: []
-        },
-        {
-          tagName: 'DIV',
-          classList: ['container']
-        },
-        {
-          tagName: 'OUTER-COMPONENT',
-          classList: []
-        },
-        {
-          tagName: 'DIV',
-          classList: ['wrapper']
-        },
-        {
-          tagName: 'INNER-COMPONENT',
-          classList: []
-        },
-        {
-          tagName: 'SPAN',
-          classList: []
-        },
-        {
-          tagName: 'svg',
-          classList: []
-        },
-        {
-          tagName: 'defs',
-          classList: []
-        },
-        {
-          tagName: 'linearGradient',
-          classList: []
-        },
-        {
-          tagName: 'stop',
-          classList: []
-        },
-        {
-          tagName: 'stop',
-          classList: []
-        },
-        {
-          tagName: 'circle',
-          classList: []
-        },
-        {
-          tagName: 'SCRIPT',
-          classList: []
-        }
-      ]
-      withDom(tagNameShadow1, context => {
-        assertResultEqual(tagName, getElementsByTagName(tagName, context), expected, true)
-      })
-    });
-
-    it('light DOM - is case insensitive', () => {
-      const tagName = "SpAn";
-      const expected = [
-        {
-          tagName: 'SPAN',
-          classList: []
-        }
-      ]
-      withDom(tagNameLight1, context => {
-        assertResultEqual(tagName, context.getElementsByTagName(tagName), expected, true)
-      })
-    });
-
-    it('shadow DOM - is case insensitive', () => {
-      const tagName = "SpAn";
-      const expected = [
-        {
-          tagName: 'SPAN',
-          classList: []
-        }
-      ]
-      withDom(tagNameShadow1, context => {
-        assertResultEqual(tagName, getElementsByTagName(tagName, context), expected, true)
-      })
-    });
-  })
-
-  describe('getElementsByTagNameNS', () => {
-    it('light DOM - returns empty with falsy ns', () => {
-      
-    });
-
-    it('shadow DOM - returns empty with falsy ns', () => {
-      
-    });
-
-    it('light DOM - returns ns matches', () => {
-      const tagName = '*';
-      const namespace = 'http://www.w3.org/2000/svg';
-      const expected = [
-        {
-          tagName: 'svg',
-          classList: []
-        },
-        {
-          tagName: 'defs',
-          classList: []
-        },
-        {
-          tagName: 'linearGradient',
-          classList: []
-        },
-        {
-          tagName: 'stop',
-          classList: []
-        },
-        {
-          tagName: 'stop',
-          classList: []
-        },
-        {
-          tagName: 'circle',
-          classList: []
-        }
-      ]
-      withDom(tagNameLight1, context => {
-        assertResultEqual(tagName, context.getElementsByTagNameNS(namespace, tagName), expected, true)
-      })
-    });
-
-    it('shadow DOM - returns ns matches', () => {
-      const tagName = '*';
-      const namespace = 'http://www.w3.org/2000/svg';
-      const expected = [
-        {
-          tagName: 'svg',
-          classList: []
-        },
-        {
-          tagName: 'defs',
-          classList: []
-        },
-        {
-          tagName: 'linearGradient',
-          classList: []
-        },
-        {
-          tagName: 'stop',
-          classList: []
-        },
-        {
-          tagName: 'stop',
-          classList: []
-        },
-        {
-          tagName: 'circle',
-          classList: []
-        }
-      ]
-      withDom(tagNameShadow1, context => {
-        assertResultEqual(tagName, getElementsByTagNameNS(namespace, tagName, context), expected, true)
-      })
-    });
-
-    it('light DOM - wildcard ns', () => {
-      const tagName = 'span';
-      const namespace = '*';
-      const expected = [
-        {
-          tagName: 'SPAN',
-          classList: []
-        }
-      ]
-      withDom(tagNameLight1, context => {
-        assertResultEqual(tagName, context.getElementsByTagNameNS(namespace, tagName), expected, true)
-      })
-    });
-
-    it('shadow DOM - wildcard ns', () => {
-      const tagName = 'span';
-      const namespace = '*';
-      const expected = [
-        {
-          tagName: 'SPAN',
-          classList: []
-        }
-      ]
-      withDom(tagNameShadow1, context => {
-        assertResultEqual(tagName, getElementsByTagNameNS(namespace, tagName, context), expected, true)
-      })
-    });
-
-    it('light DOM - wildcard tagName', () => {
-      const tagName = "*";
-      const expected = [
-        {
-          tagName: 'HTML',
-          classList: []
-        },
-        {
-          tagName: 'HEAD',
-          classList: []
-        },
-        {
-          tagName: 'BODY',
-          classList: []
-        },
-        {
-          tagName: 'DIV',
-          classList: ['container']
-        },
-        {
-          tagName: 'DIV',
-          classList: ['outer-component']
-        },
-        {
-          tagName: 'DIV',
-          classList: ['wrapper']
-        },
-        {
-          tagName: 'DIV',
-          classList: ['inner-component']
-        },
-        {
-          tagName: 'SPAN',
-          classList: []
-        }
-      ]
-      withDom(tagNameLight1, context => {
-        assertResultEqual(tagName, context.getElementsByTagNameNS('http://www.w3.org/1999/xhtml', tagName), expected, true)
-      })
-    });
-
-    it('shadow DOM - wildcard tagName', () => {
-      const tagName = "*";
-      const expected = [
-        {
-          tagName: 'HEAD',
-          classList: []
-        },
-        {
-          tagName: 'BODY',
-          classList: []
-        },
-        {
-          tagName: 'DIV',
-          classList: ['container']
-        },
-        {
-          tagName: 'OUTER-COMPONENT',
-          classList: []
-        },
-        {
-          tagName: 'DIV',
-          classList: ['wrapper']
-        },
-        {
-          tagName: 'INNER-COMPONENT',
-          classList: []
-        },
-        {
-          tagName: 'SPAN',
-          classList: []
-        },
-        {
-          tagName: 'SCRIPT',
-          classList: []
-        }
-      ]
-      withDom(tagNameShadow1, context => {
-        assertResultEqual(tagName, getElementsByTagNameNS('http://www.w3.org/1999/xhtml', tagName, context), expected, true)
-      })
-    });
   })
 })
